# openapi.yaml
openapi: 3.0.3
info:
  title: Omniscience API
  description: REST API for managing educational data including courses, classes, students, and attendance records
  version: 1.0.0
  contact:
    name: Omniscience API Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Supabase authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    EvaluationCategory:
      type: object
      properties:
        id:
          type: integer
          format: uint
        course_id:
          type: integer
          format: uint
        name:
          type: string
        description:
          type: string
        weight:
          type: number
          format: float
          description: Weight of this category in final grade calculation (percentage)
        max_score:
          type: number
          format: float
          description: Maximum score for items in this category
        drop_lowest:
          type: integer
          description: Number of lowest scores to drop from this category
        is_extra_credit:
          type: boolean
          description: Whether this category is for extra credit
        display_order:
          type: integer
          description: Order in which to display this category
        is_active:
          type: boolean
          description: Whether this category is currently active
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - course_id
        - name
        - weight
    EvaluationCategoryRequest:
      type: object
      properties:
        course_id:
          type: integer
          format: uint
        name:
          type: string
        description:
          type: string
        weight:
          type: number
          format: float
        max_score:
          type: number
          format: float
        drop_lowest:
          type: integer
          default: 0
        is_extra_credit:
          type: boolean
          default: false
        display_order:
          type: integer
      required:
        - course_id
        - name
        - weight
...
    EvaluationItem:
      type: object
      properties:
        id:
          type: integer
          format: uint
        category_id:
          type: integer
          format: uint
        student_class_id:
          type: integer
          format: uint
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        due_date:
          type: string
          format: date-time
        max_score:
          type: number
          format: float
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - category_id
        - student_class_id
        - name
        - max_score
    EvaluationItemRequest:
      type: object
      properties:
        category_id:
          type: integer
          format: uint
        student_class_id:
          type: integer
          format: uint
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        due_date:
          type: string
          format: date-time
        max_score:
          type: number
          format: float
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          default: DRAFT
      required:
        - category_id
        - student_class_id
        - name
        - max_score
...
security:
  - bearerAuth: []
paths:
...
  /evaluation-categories:
    get:
      summary: List evaluation categories
      description: Lists all evaluation categories for a course
      operationId: listEvaluationCategories
      tags:
        - Evaluation Categories
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: query
          description: ID of the course
          required: true
          schema:
            type: integer
            format: uint
        - name: include_inactive
          in: query
          description: Include inactive categories
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationCategory'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create evaluation category
      description: Create a new evaluation category
      operationId: createEvaluationCategory
      tags:
        - Evaluation Categories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationCategoryRequest'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationCategory'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /evaluation-categories/{id}:
    get:
      summary: Get evaluation category
      description: Get a specific evaluation category by ID
      operationId: getEvaluationCategory
      tags:
        - Evaluation Categories
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationCategory'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update evaluation category
      description: Update an existing evaluation category
      operationId: updateEvaluationCategory
      tags:
        - Evaluation Categories
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationCategoryRequest'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationCategory'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete evaluation category
      description: Delete an evaluation category (cascades to evaluation items and grades)
      operationId: deleteEvaluationCategory
      tags:
        - Evaluation Categories
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '204':
          description: Category deleted successfully
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /evaluation-items:
    get:
      summary: List evaluation items
      description: List evaluation items (e.g., Quiz 1, Midterm Exam)
      operationId: listEvaluationItems
      tags:
        - Evaluation Items
      security:
        - bearerAuth: []
      parameters:
        - name: category_id
          in: query
          description: Filter by category
          required: false
          schema:
            type: integer
            format: uint
        - name: student_class_id
          in: query
          description: Filter by class
          required: false
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationItem'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create evaluation item
      description: Create a new evaluation item
      operationId: createEvaluationItem
      tags:
        - Evaluation Items
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationItem'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /evaluation-items/{id}:
    get:
      summary: Get evaluation item
      description: Get a specific evaluation item
      operationId: getEvaluationItem
      tags:
        - Evaluation Items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Item ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationItem'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update evaluation item
      description: Update an evaluation item
      operationId: updateEvaluationItem
      tags:
        - Evaluation Items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Item ID
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationItemRequest'
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationItem'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete evaluation item
      description: Delete an evaluation item (cascades to grades)
      operationId: deleteEvaluationItem
      tags:
        - Evaluation Items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Item ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '204':
          description: Item deleted successfully
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...